Get Monitoring Location Identifier and Assessment Unit Identifier Crosswalk from ATTAINS

Tribes and States who participate in electronic reporting of water quality conditions through EPA ATTAINS may optionally submit a crosswalk of WQP monitoring location identifiers associated with their assessment units to ATTAINS. If an organization has recorded this information in ATTAINS, this function can be used to get the ATTAINS user submitted crosswalk of known monitoring location identifiers and assessment units. As of 2025, all tribal nations record this information in ATTAINS but only a few states.

Usage

TADA_GetATTAINSAUSiteCrosswalk(org_id = NULL)

Arguments

org_id

The ATTAINS organization identifier must be supplied by the user. A list of organization identifiers can be found by downloading the ATTAINS Domains Excel file: https://www.epa.gov/system/files/other-files/2023-09/DOMAINS.xlsx. Organization identifiers are listed in the "OrgName" tab. The "code" column contains the organization identifiers that should be used for this param.

Value

A dataframe with four columns, MonitoringLocationIdentifier, OrganizationIdentifier, ATTAINS.assessmentunitidentifier, and MonitoringDataLinkText is returned. This is the crosswalk between monitoring location identifiers and assessment units that the state or tribal organization submitted to ATTAINS (optional). If an ATTAINS organization has not submitted this information in ATTAINS, the function will not return a dataframe.

Examples

if (FALSE) { # \dontrun{
# Alaska example
AK_crosswalk <- TADA_GetATTAINSAUSiteCrosswalk(org_id = "AKDECWQ")

# Pueblo of Tesuque example
PUEBLOOFTESUQUE_crosswalk <- TADA_GetATTAINSAUSiteCrosswalk(
  org_id = "PUEBLOOFTESUQUE"
)

# Arizona example, returns blank dataframe as of 1/21/25
AZ_crosswalk <- TADA_GetATTAINSAUSiteCrosswalk(org_id = "21ARIZ")
} # }