Create Scatterplot(s)

Create Scatterplot(s)

Usage

TADA_Scatterplot(.data, id_cols = c("TADA.ComparableDataIdentifier"))

Arguments

.data

TADA dataframe containing the data downloaded from the WQP, where each row represents a unique data record. Dataframe must include the columns 'TADA.ComparableDataIdentifier', 'TADA.ResultMeasureValue', and 'TADA.ResultMeasure.MeasureUnitCode' to run this function. The 'TADA.ComparableDataIdentifier' column is added to the dataframe automatically when WQP data is retrieved using TADADataRetrieval. This TADA.ComparableDataIdentifier can be updated to harmonize synonyms by running the function TADA_HarmonizeSynonyms(). You can also include additional grouping columns in the id_cols input if desired. If more than one TADA.ComparableDataIdentifier exists in the dataset, the function creates a list of plots, where each list element name is a unique TADA.ComparableDataIdentifier.

id_cols

The column(s) in the dataset used to identify the unique groups to be plotted. Defaults to 'TADA.ComparableDataIdentifier'.

Value

A list of plotly scatterplot figures showing the distribution of sample values for each comparable data group (TADA.ComparableDataIdentifier).

Examples

# Load example dataset:
utils::data(Data_TribalNations_Harmonized)

# Create a scatterplot for each comparable data group (TADA.ComparableDataIdentifier)
# in the input dataframe:
TADA_Scatterplot(Data_TribalNations_Harmonized, id_cols = "TADA.ComparableDataIdentifier")
#> $`ALKALINITY, BICARBONATE AS CACO3 UG/L`
#> 
#> $`ALKALINITY, CARBONATE AS CACO3 UG/L`
#> 
#> $`ALKALINITY, TOTAL AS CACO3 UG/L`
#> 
#> $`ALKALINITY TOTAL AS CACO3 UG/L`
#> 
#> $`AMMONIA FILTERED AS N MG/L`
#> 
#> $`AMMONIA UNFILTERED AS N MG/L`
#> 
#> $`ANATOXIN A UG/L`
#> 
#> $`CADMIUM TOTAL RECOVERABLE UG/L`
#> 
#> $`CADMIUM TOTAL UG/L`
#> 
#> $`CHLOROPHYLL A, CORRECTED FOR PHEOPHYTIN SUSPENDED UG/L`
#> 
#> $`CHLOROPHYLL A FILTERED AS CHLOROPHYLL A G/M2`
#> 
#> $`CHLOROPHYLL A FILTERED AS CHLOROPHYLL A UG/L`
#> 
#> $`CHLOROPHYLL A UNFILTERED AS CHLOROPHYLL A UG/L`
#> 
#> $`CHROMIUM DISSOLVED UG/L`
#> 
#> $`CHROMIUM TOTAL RECOVERABLE UG/L`
#> 
#> $`CHROMIUM TOTAL UG/L`
#> 
#> $`CONDITION CLASS (DISSOLVED OXYGEN (DO)) %`
#> 
#> $`COPPER DISSOLVED UG/L`
#> 
#> $`COPPER TOTAL RECOVERABLE UG/L`
#> 
#> $`COPPER TOTAL UG/L`
#> 
#> $`DENSITY UG/L`
#> 
#> $`DEPTH, SECCHI DISK DEPTH M`
#> 
#> $`DEPTH, SNOW COVER IN`
#> 
#> $`DEPTH M`
#> 
#> $`DIBROMOFLUOROMETHANE UG/L`
#> 
#> $`DISCHARGE, RIVER/STREAM CFS`
#> 
#> $`DISSOLVED OXYGEN (DO) MG/L`
#> 
#> $`DISSOLVED OXYGEN SATURATION %`
#> 
#> $`DISSOLVED OXYGEN UPTAKE UG/L`
#> 
#> $`ESCHERICHIA COLI CFU/100ML`
#> 
#> $`EXTRACTABLE ORGANIC MATTER (EOM) %`
#> 
#> $`EXTRACTABLE ORGANIC MATTER (EOM) G/M2`
#> 
#> $`HARDNESS, CARBONATE AS CACO3 UG/L`
#> 
#> $`HARDNESS, CARBONATE UG/L`
#> 
#> $`HEIGHT, GAGE M`
#> 
#> $`IRON DISSOLVED UG/L`
#> 
#> $`IRON TOTAL RECOVERABLE UG/L`
#> 
#> $`LEAD DISSOLVED UG/L`
#> 
#> $`LEAD TOTAL RECOVERABLE UG/L`
#> 
#> $`LEAD TOTAL UG/L`
#> 
#> $`LOONS, VISUAL OBSERVATION COUNT`
#> 
#> $`MAGNESIUM DISSOLVED UG/L`
#> 
#> $`MAGNESIUM TOTAL RECOVERABLE UG/L`
#> 
#> $`MERCURY DISSOLVED UG/L`
#> 
#> $`MERCURY FILTERED, LAB UG/L`
#> 
#> $`MERCURY TOTAL UG/L`
#> 
#> $`MERCURY UNFILTERED UG/L`
#> 
#> $`NICKEL DISSOLVED UG/L`
#> 
#> $`NICKEL TOTAL RECOVERABLE UG/L`
#> 
#> $`NICKEL TOTAL UG/L`
#> 
#> $`NITRATE + NITRITE FILTERED AS N MG/L`
#> 
#> $`NITRATE + NITRITE UNFILTERED AS N MG/L`
#> 
#> $`NITRATE UNFILTERED AS N MG/L`
#> 
#> $`OCTADECANE, 1-CHLORO- TOTAL RECOVERABLE UG/L`
#> 
#> $`OCTADECANE, 1-CHLORO- TOTAL UG/L`
#> 
#> $`OCTANE, 1-CHLORO- UG/L`
#> 
#> $`ORGANIC CARBON DISSOLVED UG/L`
#> 
#> $`ORGANIC CARBON TOTAL UG/L`
#> 
#> $`ORTHOPHOSPHATE FILTERED AS P UG/L`
#> 
#> $`ORTHOPHOSPHATE UNFILTERED AS P UG/L`
#> 
#> $`P-BROMOFLUOROBENZENE TOTAL RECOVERABLE UG/L`
#> 
#> $`P-BROMOFLUOROBENZENE TOTAL UG/L`
#> 
#> $`PARTIAL PRESSURE O2 G/M2`
#> 
#> $`PERIPHYTON G/M2`
#> 
#> $PH
#> 
#> $`PRESSURE G/M2`
#> 
#> $`RADIUM-228 PCI/L`
#> 
#> $`RESISTIVITY US/CM`
#> 
#> $`SALINITY UG/KG`
#> 
#> $`SILICA DISSOLVED AS SIO2 UG/L`
#> 
#> $`SILICA AS SIO2 UG/L`
#> 
#> $`SULFATE TOTAL AS SO4 UG/L`
#> 
#> $`TEMPERATURE DEG C`
#> 
#> $`TOLUENE-D8 TOTAL RECOVERABLE UG/L`
#> 
#> $`TOLUENE-D8 TOTAL UG/L`
#> 
#> $`TOTAL HARDNESS TOTAL AS CACO3 MG/L`
#> 
#> $`TOTAL KJELDAHL NITROGEN (ORGANIC N & NH3) UNFILTERED AS N MG/L`
#> 
#> $`TOTAL MICROCYSTINS PLUS NODULARINS UG/L`
#> 
#> $`TOTAL NITROGEN, MIXED FORMS UNFILTERED AS N MG/L`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS FILTERED AS P UG/L`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L`
#> 
#> $`TOTAL SUSPENDED SOLIDS NON-FILTERABLE (PARTICLE) UG/L`
#> 
#> $`TRANSPARENCY, SECCHI TUBE WITH DISK IN`
#> 
#> $`TURBIDITY NTU`
#> 
#> $`VOLATILE SUSPENDED SOLIDS TOTAL UG/L`
#> 
#> $`WIND SPEED SEVERITY (CHOICE LIST)`
#> 
#> $`WIND VELOCITY M/SEC`
#> 
#> $`ZINC DISSOLVED UG/L`
#> 
#> $`ZINC TOTAL RECOVERABLE UG/L`
#> 
#> $`ZINC TOTAL UG/L`
#> 

# Create a single scatterplot using defaults. The input dataframe in this
# example is filtered so it includes only one TADA.ComparableDataIdentifier
df <- dplyr::filter(
  Data_TribalNations_Harmonized,
  TADA.ComparableDataIdentifier ==
    "TOTAL PHOSPHORUS, MIXED FORMS_UNFILTERED_AS P_UG/L"
)
TADA_Scatterplot(df, id_cols = "TADA.ComparableDataIdentifier")

# Creates a scatterplot for each monitoring location
TADA_Scatterplot(df, id_cols = c("TADA.ComparableDataIdentifier", "MonitoringLocationName"))
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Arnoux Creek`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Aspen Ranch`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Badger Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Badger Creek #4`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Bang (Long) Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Barker Arroyo`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Bass Lake: Northwest Basin`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Bass Lake: Southeast Basin`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Battle River Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Battle River North Branch`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Battle River South Branch`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Bear Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Big Lake (North Basin)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Big Lake (South Basin)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Big Rock Coulee`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Big Thunder Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Big Traverse Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Birch Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Birch Creek #2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Birch Creek #3`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Birch Creek #4 (aka Two Med)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Birch Creek #5`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Birch Creek #6`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Birch Two Medicine #10`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Blackduck River Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Blackduck River at Hwy 1 Bridge`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Blacktail Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Blacktail Creek #2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Boulder Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Buffalo Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cedar Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Chain: Middle Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Chain: South Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Clearwater River CLWR-B`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Clearwater River at Boundary`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cormorant River at Hwy 23 Culverts`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cottonwood Wash Highway 95`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cottonwood Wash Spring`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cottonwood wash south 2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cut Bank Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cut Bank Creek #3`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cut Bank Creek #4`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cut Bank Creek #5`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cut Bank Creek #6`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cut Bank Creek #8`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Cut Bank Creek #9`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Deadfish Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Depot Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Dickens Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Divide Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Dry Fork Milk River #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Emerald Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Fond du Lac Creek Station 2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Fox Creek #2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Fullers Lake: East Basin`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Fullers Lake: West Basin`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Greasewood #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Green Lake (Red Lake)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Green Lake (Redby)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Hanna Spring`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Harrison Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Hay Creek on Ponemah Rd`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Heart Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Infiltration Diversion/Upper Ponds`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Island Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Jaskari Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Joe Martin Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L John Weathered`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Johnson Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Kennedy Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Kinney Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L LOW: MPCA Co-located Site Big Traverse Bay Site 5`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L LOW: MPCA Co-located Site Four Mile Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L LOW: MPCA Co-located Site Muskeg Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L LOW: MPCA Co-located Site Near Long Point`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lac Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lake of the Woods at Big Traverse Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lake of the Woods at Central Big Traverse Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lake of the Woods at Muskeg Bay`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Little Badger Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Little Rock Coulee`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Little Rock Creek at East Fireline Rd`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Little Rock Creek on Shell Lake Rd`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Little Thunder Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Livermore Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Long Lake (Burt)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lost Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lower Red Lake : Central`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lower Red Lake : East`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lower Red Lake : East Central`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lower Red Lake : West`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Lower Red Lake: West Central`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mancos River at State Line near 4 Corners`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mancos River at gauging station upstream from Hwy. 666`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Martin Branch (Marshall Rd.)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mc Elmo Creek @ J Bridge`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L McElmo creek at Ute creek CON.`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Middle Fork Milk River #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Midvale Creek #3`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Milk River #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Milk River #2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Milk River #3`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Milk River #4`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Milk River #5`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Miller (Mud) Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Morrison Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mud River Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mud River Middle`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mud River at MUDR-M`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mud River at MUDR-U`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mud River at South Boundary Road`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Mud River on East Fireline Trail`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Navajo Pond`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Navajo Wash above CR A on Don Schwindt''s land`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Navajo Wash at gauging station #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L No-Name Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L North Fork Cut Bank Creek #3`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L North Fork Milk River #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L North Fork Milk River #2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pat Martin Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Perch Lake (North Basin)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Perch Lake (South Basin)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pickerel Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek Off Reservation`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek at Barton's Camp Trail`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek at PIKE-B`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek at PIKE-I`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek at South Boundary`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Pike Creek on West Fireline Rd`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Poplar Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Poplar Creek by Upper Red Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Railroad Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Red Lake River at Outlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Red Lake River at Rock Dam`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Rice Portage Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Rio Chupadero`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Rio Tesuque`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Round Lake (Sylvia)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L SanJuan @ USGS Guaging station`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Sandy Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Sandy River Upstream`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Sandy River at Gonvick Truck Trail`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Sandy River at SANR-U`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Second Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Shell Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Shemahgun Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Simian Creek Station 1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Simian Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Sofie Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L South Fork Cut Bank Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L South Fork Milk River #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L South Fork Milk River #2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L South Fork Two Medicine River #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L South Fork Whitetail Creek #2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Spruce (Spirit) Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Squaw Smith Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L St. Mary River #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L St. Mary River #2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L St. Mary River #4`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L St. Mary River #5`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Stoney Brook Station 2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Stony Creek Inlet`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Swiftcurrent Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Swiftcurrent Creek #2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L THIRD LAKE (16')`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L THIRD LAKE (18')`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L THIRD LAKE (19')`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Tamarac River at Steel Bridge`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Third Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Toad Creek #2`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Two Medicine River #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Two Medicine River #10`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Two Medicine River #3`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Two Medicine River #5`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Two Medicine River #6`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Upper Red Lake: Central`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Upper Red Lake: East`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Upper Red Lake: East Central`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Upper Red Lake: West`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Upper Red Lake: West-Central`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L West Twin Lake (North Basin)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L West Twin Lake (South Basin)`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Whitetail Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Wild Rice Lake`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Willow Creek #1`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Willow Creek #3`
#> 
#> $`TOTAL PHOSPHORUS, MIXED FORMS UNFILTERED AS P UG/L Willow Creek #5`
#> 

# Create multiple scatterplots with additional grouping columns and view the first
# plot in list. In this example, we will group by both TADA.ComparableDataIdentifier
# and MonitoringLocationTypeName (e.g. stream, reservoir, canal, etc.)
# Load example dataset:
utils::data(Data_Nutrients_UT)
Scatterplot_output <- TADA_Scatterplot(Data_Nutrients_UT,
  id_cols = c("TADA.ComparableDataIdentifier", "MonitoringLocationTypeName")
)
# This example generates 47 scatterplots
Scatterplot_output[[10]]

Scatterplot_output[[25]]

Scatterplot_output[[35]]