addAllATTAINS Internal function to add all ATTAINS assessment units (lines, points, or polygons) or ATTAINS catchments (polygons) to TADA maps.

addAllATTAINS Internal function to add all ATTAINS assessment units (lines, points, or polygons) or ATTAINS catchments (polygons) to TADA maps.

Usage

addAllATTAINS(
  .data = NULL,
  map = NULL,
  points_layer = NULL,
  polygons_layer = NULL,
  lines_layer = NULL,
  catchment_layer = NULL,
  outline_layer = NULL,
  missing_raw_layer = NULL,
  overlay_groups = NULL,
  icons = NULL,
  catchment = NULL
)

Arguments

.data

A data frame created using prepATTAINSMapper (must contain a geometry column)

map

The TADA leaflet map to add the ATTAINS geometry to. Required, function will not run with map is missing or NULL. Default is map = NULL.

points_layer

Optional data frame argument. Contains the data required to map ATTAINS assessment unit point geometry. Default = NULL.

polygons_layer

Optional data frame argument. Contains the data required to map ATTAINS assessment unit polygon geometry. When polygons_layer = NULL, the polygon assessment units data are not used when searching for missing raw features. Default = NULL.

lines_layer

Optional data frame argument. Contains the data required to map ATTAINS assessment unit line geometry. When lines_layer = NULL, the line assessment units data are not used when searching for missing raw features. Default = NULL.

catchment_layer

Optional data frame argument. Contains the data required to map ATTAINS catchment geometry. Default = NULL.

outline_layer

Optional data frame argument. Contains the data required to map catchments for monitoring locations not associated with an assessment unit. Default = NULL.

missing_raw_layer

Optional data frame argument. Contains the data required to map ATTAINS catchments for ATTAINS assessment units that are missing geometry. Default = NULL.

overlay_groups

Initialized vector to add names of groups to map. This is to allow users to toggle specific layers on/off. If it is NULL, the function will fail with an error message. Default is overlay_list = NULL.

icons

Character argument. The list of icon paths generated by the internal function getMapIconLabels. If already called in a larger mapping function, it can be referenced here (for efficiency). If icons = NULL, getMapIconLabels will run and fetch the list. Default is icons = NULL. This argument is only applied to for point assessment units.

catchment

Boolean argument. Specifies whether the geometries being mapped are catchments or assessment units to determine color and other attributes. Default is catchment = FALSE, meaning that assessment units are being mapped.

Value

ATTAINS geometry correctly formatted for display in a TADA leaflet map.